Toast Frame: Make white background optional for android 15

+++ app/src/main/java/com/drdisagree/iconify/utils/overlay/compiler/OnDemandCompiler.kt
@@ -1,6 +1,7 @@
 package com.drdisagree.iconify.utils.overlay.compiler
 
 import android.util.Log
+import com.drdisagree.iconify.data.common.Preferences.TOAST_FRAME_ALWAYS_WHITE
 import com.drdisagree.iconify.data.common.Resources
 import com.drdisagree.iconify.data.common.Resources.BACKUP_DIR
 import com.drdisagree.iconify.data.common.Resources.DATA_DIR
@@ -11,6 +12,7 @@ import com.drdisagree.iconify.data.common.Resources.TEMP_DIR
 import com.drdisagree.iconify.data.common.Resources.TEMP_OVERLAY_DIR
 import com.drdisagree.iconify.data.common.Resources.UNSIGNED_DIR
 import com.drdisagree.iconify.data.common.Resources.UNSIGNED_UNALIGNED_DIR
+import com.drdisagree.iconify.data.config.RPrefs.getBoolean
 import com.drdisagree.iconify.utils.FileUtils.copyAssets
 import com.drdisagree.iconify.utils.MiscUtils.requiresNewToastStyle
 import com.drdisagree.iconify.utils.RootUtils.setPermissions
@@ -172,7 +174,10 @@ object OnDemandCompiler {
     }
 
     private fun handleNewToastStyle() {
-        if (mOverlayName != "TSTFRM" || !requiresNewToastStyle()) return
+        if (mOverlayName != "TSTFRM" ||
+            !requiresNewToastStyle() ||
+            !getBoolean(TOAST_FRAME_ALWAYS_WHITE, false)
+        ) return
 
         Shell.cmd(
             "find \"$TEMP_CACHE_DIR/$mPackage/$mOverlayName\" -type d -name \"*-night\" -exec rm -rf {} +",
